+++ srcs/lib/ft_printf/srcs/ft_get_strlen.c
@@ -30,11 +30,11 @@ char	*ft_get_string_nb(t_pdata *data)
 
 	s = NULL;
 	if (data->convers == 'd' || data->convers == 'i')
-		s = ft_itoa((int)data->arg);
+		s = ft_itoa((long)data->arg);
 	else if (data->convers == 'u')
-		s = ft_unsigned_itoa((unsigned int)data->arg);
+		s = ft_unsigned_itoa((unsigned long)data->arg);
 	else if (data->convers == 'x' || data->convers == 'X')
-		s = ft_itox((unsigned int)data->arg, data);
+		s = ft_itox((unsigned long)data->arg, data);
 	else if (data->convers == 'p')
 		s = ft_ultox((unsigned long)data->arg);
 	else if (ft_is_convers(data->convers) == 0)
@@ -54,11 +54,11 @@ int		ft_get_strlen(t_pdata *data)
 	else if (data->convers == 'p')
 		return (ft_slen(ft_ultox((unsigned long)data->arg)) + 2);
 	else if (data->convers == 'd' || data->convers == 'i')
-		return (ft_slen(ft_itoa((int)data->arg)));
+		return (ft_slen(ft_itoa((long)data->arg)));
 	else if (data->convers == 'u')
-		return (ft_slen(ft_unsigned_itoa((int)data->arg)));
+		return (ft_slen(ft_unsigned_itoa((unsigned long)data->arg)));
 	else if (data->convers == 'x' || data->convers == 'X')
-		return (ft_slen(ft_itox((unsigned int)data->arg, data)));
+		return (ft_slen(ft_itox((unsigned long)data->arg, data)));
 	else if (ft_is_convers(data->convers) == 0)
 		return (1);
 	return (0);
